Add status enum, general cleaning and bug fixes

+++ src/user.h
@@ -2,7 +2,7 @@
 #define USER_H
 
 struct user {
-  uint16_t uid;
+  int uid;
   char handle[HANDLE_LEN + 1];
   // XXX friends
 };
